Interactively open/run sets of smoke tests  🐿 v2.6.0
