+++ README.md
@@ -15,6 +15,10 @@ Rather than the chore of writing tests, simply add some JS config (default locat
 
 `n-test smoke --config path/to/config.js --host https://local.ft.com:3002`
 
+`n-test smoke basic` - runs just the set with the name 'basic'
+
+`n-test smoke -i` - interactively select the suites to run
+
 *Example config*
 ```
 module.exports = [
@@ -82,14 +86,16 @@ module.exports = [
 const nTest = require('@financial-times/n-test');
 nTest.smoke.run({ auth: true, host: 'local.ft.com:3002' })
 	.then((results) => { //all passed })
-	.catch((results) => { //some failed })
+	.catch((results) => { //some failed });
+
+nTest.smoke.run({}, ['basic']);
 ```
 
 #### Open
 
 Opens an instance of Chromium with all of the URLs specified in the smoke tests, for manual verification.
 
-`n-test open`
+`n-test open` - interactively select which sets of of URLs to open
 
 `n-test open headers --breakpoint M --config path/to/config.js --host https://local.ft.com:3002`
