+++ lib/smoke/index.js
@@ -2,8 +2,9 @@ const path = require('path');
 const fs = require('fs');
 
 const test = require('./test');
+const filterConfigs = require('./filter-configs');
 
-module.exports.run = (opts) => {
+module.exports.run = async (opts, sets) => {
 	const configFile = path.join(process.cwd(), opts.config || 'test/smoke.js');
 	if (!fs.existsSync(configFile)) {
 		throw new Error(`Config file for smoke test does not exist at ${configFile}. Either create a config file at ./test/smoke.js, or pass in a path using --config.`);
@@ -19,6 +20,8 @@ module.exports.run = (opts) => {
 			host = 'http://' + host;
 		}
 
-		return test(configFile, host, opts.auth);
+		const configsToRun = await filterConfigs(configFile, sets, opts.interactive);
+
+		return test(configsToRun, host, opts.auth);
 	}
 };
