+++ lib/smoke/open.js
@@ -2,43 +2,19 @@ const path = require('path');
 const fs = require('fs');
 const puppeteer = require('puppeteer');
 const TestPage = require('./test-page');
-const inquirer = require('inquirer');
-
-const askForSets = async (config) => {
-	const possibilities = config.map((test, index)=> test.name || 'Untitled ' + index);
-	const responses = await inquirer.prompt({
-		name: 'choices',
-		type: 'checkbox',
-		message: 'Select the URL sets to open',
-		choices: possibilities
-	});
-	return [].concat(responses.choices);
-};
+const filterConfigs = require('./filter-configs');
 
 
-module.exports = async (sets, opts) => {
+
+module.exports = async (opts, sets) => {
 	const configFile = path.join(process.cwd(), opts.config || 'test/smoke.js');
 	if (!fs.existsSync(configFile)) {
 		throw new Error(`Config file for smoke test does not exist at ${configFile}. Either create a config file at ./test/smoke.js, or pass in a path using --config.`);
 	} else {
-		const config = require(configFile);
-		let configsToOpen;
-		if(!(sets && sets.length)) {
-			sets = await askForSets(config);
-		}
-
 
 		const host = opts.host || 'http://localhost:3002';
-		const browser = await puppeteer.launch({headless: false});
-
-		if(sets && sets.length) {
-			configsToOpen = config.filter((test, index) => (sets.includes(test.name) || sets.includes('Untitled ' + index)));
-		} else {
-			configsToOpen = config;
-		}
-
-		// eslint-disable-next-line no-console
-		console.info('Opening URLs from test sets: ', sets);
+		const browser = await puppeteer.launch({ headless: false });
+		const configsToOpen = await filterConfigs(configFile, sets, true);
 
 		configsToOpen.forEach((suiteOpts) => {
 			Object.keys(suiteOpts.urls).forEach(async (url) => {
@@ -51,7 +27,7 @@ module.exports = async (sets, opts) => {
 				urlOpts.method = urlOpts.method || suiteOpts.method;
 				urlOpts.body = urlOpts.body || suiteOpts.body;
 
-				urlOpts.breakpoint = urlOpts.breakpoint || suiteOpts.breakpoint;
+				urlOpts.breakpoint = urlOpts.breakpoint || suiteOpts.breakpoint || opts.breakpoint;
 
 				const fullUrl = `${host}${url}`;
 				const testPage = new TestPage(fullUrl, urlOpts);
