+++ lib/smoke/test.js
@@ -4,8 +4,7 @@ const verifyUrl = require('./verify-url');
 
 const directly = require('directly');
 
-const run = async (configFile, host, authenticate) => {
-	const config = require(configFile);
+const run = async (config, host, authenticate) => {
 	const browser = await puppeteer.launch();
 
 	const urlTests = [];
